+++ exploreWithMe/src/main/java/ru/practicum/exploreWithMe/service/event/EventAdminService.java
@@ -7,8 +7,8 @@ import ru.practicum.exploreWithMe.enums.State;
 import java.util.List;
 
 public interface EventAdminService {
-    List<EventFullDto> getEvents(List<Integer> users, List<State> states,
-                                 List<Integer> categories, String rangeStart,
+    List<EventFullDto> getEvents(List<Long> users, List<State> states,
+                                 List<Long> categories, String rangeStart,
                                  String rangeEnd, int from, int size);
 
     EventFullDto updateEvent(Long eventId, UpdateEventRequest updateEventRequest);
